Update Discord link in the README
The discord community link appears twice in the README, once on top (was already fixed through a separated issue that I've opened previously). The second place is in line 102 under the "Community" title. This commit updates the "second" discord link with the updated valid link from the previous issue that was already fixed.